+++ gDRplots/R/plot-normalization_qc.R
@@ -1,11 +1,12 @@
-#' Plot violin for normalized single-agent data to check quality of data
+#' Plot violin for normalized or averaged single-agent data to control quality of the data
 #'
 #' @param dt_assay data.table representation of the data in assay
-#'    output from \code{gDRutils::convert_se_assay_to_dt(se, <assya_name>)}
+#'    output from \code{gDRutils::convert_se_assay_to_dt(se, <assay_name>)} 
+#'    for assay_name like "Normalized" and "Averaged"
 #' @param cl_name string cell line name to be plotted (Cell Line Name)
 #' @param metric string with variable name to be plotted; it has to be in \code{dt_assay}
-#' @param metric_growth string with normalization_types to be selected
-#'                      one of: "GR" ("GRvalue") or "RV" ("RelativeViability")
+#' @param normalization_type string with normalization_types to be selected
+#'                           one of: "GR" ("GRvalue") or "RV" ("RelativeViability")
 #'
 #' @return plot with violin for each drug
 #'
@@ -23,31 +24,36 @@
 #' dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
 #' plot_var_distribution_qc(dt_assay = dt_average,
 #'                          cl_name = cl_name,
-#'                          metric_growth = "RV")
-#' 
-#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
-#' plot_var_distribution_qc(dt_assay = dt_metrics,
+#'                          normalization_type = "RV")
+#'                          
+#' plot_var_distribution_qc(dt_assay = dt_average,
 #'                          cl_name = cl_name,
-#'                          metric = "r2")
-#'  
+#'                          metric = "x_std",
+#'                          normalization_type = "RV")
+#'                          
 #' @export
 plot_var_distribution_qc <- function(dt_assay,
                                      cl_name,
                                      metric = "x",
-                                     metric_growth = "GR") {
+                                     normalization_type = "GR") {
   
   checkmate::expect_data_table(dt_assay)
   checkmate::expect_string(cl_name)
   checkmate::expect_choice(metric, choices = names(dt_assay))
-  checkmate::expect_choice(metric_growth, choices = c("GR", "RV"))
+  checkmate::expect_choice(normalization_type, choices = c("GR", "RV"))
   
   cellline_name <- gDRutils::get_env_identifiers("cellline_name")
   clid <- gDRutils::get_env_identifiers("cellline")
   drug_name <- gDRutils::get_env_identifiers("drug_name")
   gnumber <- gDRutils::get_env_identifiers("drug")
   
-  cl_clid <- unique(dt_assay[get(cellline_name) == cl_name, ][[clid]]) 
-  tab_subplot <- dt_assay[normalization_type == metric_growth & get(cellline_name) == cl_name, ]
+  cl_clid <- unique(dt_assay[get(cellline_name) == cl_name, clid]) 
+  # filter data for normalization type
+  data.table::setkeyv(dt_assay, "normalization_type")
+  dt_assay <- dt_assay[normalization_type]
+  data.table::setkey(dt_assay, NULL)
+  
+  tab_subplot <- dt_assay[get(cellline_name) == cl_name, ]
   
   plt_title <- sprintf("%s (%s)", cl_name, cl_clid)
   color_palette <- get_qual_colors(NROW(unique(tab_subplot[[drug_name]])))
@@ -60,21 +66,108 @@ plot_var_distribution_qc <- function(dt_assay,
     ggplot2::theme_minimal() +
     ggplot2::scale_fill_manual(values = color_palette) +
     ggplot2::scale_color_manual(values = color_palette) +
-    ggplot2::labs(y = metric_growth, x = drug_name, title = plt_title) +
+    ggplot2::labs(y = sprintf("%s for %s", metric, normalization_type), x = drug_name, title = plt_title) +
     ggplot2::theme(legend.position = "none",
                    axis.text.x = ggplot2::element_text(angle = 45, vjust = 1, hjust = 1))
-
-  return(suppressWarnings(plt))
+  
+  return(plt)
 }
 
-#' Plot drug response curves for single-agent data to check quality of data
+#' Lollipop plot for metric single-agent data to control quality of the data
+#'
+#' @param dt_assay data.table representation of the data in assay
+#'    output from \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
+#' @param cl_name string cell line name to be plotted (Cell Line Name)
+#' @param metric string with variable name to be plotted; it has to be in \code{dt_assay}
+#' @param normalization_type string with normalization_types to be selected
+#'                           one of: "GR" ("GRvalue") or "RV" ("RelativeViability")
+#' @param with_table logical whether table with metric values should be shown next to the plot
+#'
+#' @return lollipop plot with stat value for each drug
+#'
+#' @keywords QC_plot
+#' @examples
+#' mae <- gDRutils::get_synthetic_data("small")
+#' se <- mae[[gDRutils::get_supported_experiments("sa")]]
+#' 
+#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
+#' cl_name <- dt_metrics[["CellLineName"]][1]
+#' 
+#' plot_var_stat_qc(dt_assay = dt_metrics,
+#'                  cl_name = cl_name)
+#'                          
+#' plot_var_stat_qc(dt_assay = dt_metrics,
+#'                  cl_name = cl_name,
+#'                  metric = "r2",
+#'                  normalization_type = "RV")
+#'                          
+#' plot_var_stat_qc(dt_assay = dt_metrics,
+#'                  cl_name = cl_name,
+#'                  metric = "x_AOC",
+#'                  normalization_type = "RV",
+#'                  with_table = TRUE)
+#'                          
+#' @export
+plot_var_stat_qc <- function(dt_assay,
+                             cl_name,
+                             metric = "r2",
+                             normalization_type = "GR", 
+                             with_table = FALSE) {
+  
+  checkmate::expect_data_table(dt_assay)
+  checkmate::expect_string(cl_name)
+  checkmate::expect_choice(metric, choices = names(dt_assay))
+  checkmate::expect_choice(normalization_type, choices = c("GR", "RV"))
+  checkmate::expect_flag(with_table)
+  
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  clid <- gDRutils::get_env_identifiers("cellline")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  gnumber <- gDRutils::get_env_identifiers("drug")
+  
+  cl_clid <- unique(dt_assay[get(cellline_name) == cl_name, clid]) 
+  # filter data for normalization type
+  data.table::setkeyv(dt_assay, "normalization_type")
+  dt_assay <- dt_assay[normalization_type]
+  data.table::setkey(dt_assay, NULL)
+  
+  tab_subplot <- dt_assay[get(cellline_name) == cl_name, ]
+  
+  plt_title <- sprintf("%s (%s)", cl_name, cl_clid)
+  color_palette <- get_qual_colors(NROW(unique(tab_subplot[[drug_name]])))
+  
+  plt <- ggplot2::ggplot(tab_subplot, ggplot2::aes(x = get(drug_name), y = !!rlang::sym(metric))) +
+    ggplot2::geom_hline(yintercept = c(0, 1), color = "#2c3e50", linetype = "dashed") +
+    ggplot2::geom_segment(
+      ggplot2::aes(x = get(drug_name), xend = get(drug_name), y = 0, yend = !!rlang::sym(metric))) +
+    ggplot2::geom_point(ggplot2::aes(fill = get(drug_name), color = get(drug_name)), 
+                        alpha = 0.75, size = 5, shape = 21, stroke = 1) +
+    ggplot2::theme_minimal() +
+    ggplot2::scale_fill_manual(values = color_palette) +
+    ggplot2::scale_color_manual(values = color_palette) +
+    ggplot2::labs(y = sprintf("%s for %s", metric, normalization_type), x = drug_name, title = plt_title) +
+    ggplot2::theme(legend.position = "none",
+                   axis.text.x = ggplot2::element_text(angle = 45, vjust = 1, hjust = 1))
+  
+  if (with_table) {
+    tab_metric <- ggpubr::ggtexttable(
+      tab_subplot[, .SD, .SDcols = c(drug_name, metric)][order(get(metric))], 
+      rows = NULL, theme = ggpubr::ttheme("light")) 
+    
+    plt <- ggpubr::ggarrange(plt, tab_metric, nrow = 1, widths = c(2, 1))
+  }
+  
+  return(plt)
+}
+
+#' Plot drug response curves for single-agent data to control quality of the data
 #'
 #' @param dt_treat data.table representation of the data in \code{RawTreated} assay
 #'    output from \code{gDRutils::convert_se_assay_to_dt(se, "RawTreated")}
 #' @param dt_controls data.table representation of the data in \code{Controls} assay
 #'    output from \code{gDRutils::convert_se_assay_to_dt(se, "Controls")}
 #'
-#' @return hetamap with check of mapping controls to treated
+#' @return hetamap of mapping controls to treated
 #'
 #' @keywords QC_plot
 #' @examples
@@ -116,7 +209,8 @@ heatmap_control_mapping_qc <- function(dt_treat,
   # Convert the result to a matrix format suitable for pheatmap
   result_matrix <- data.table::dcast(result, rId ~ cId, value.var = "N")
   rownames <- result_matrix$rId
-  result_matrix <- as.matrix(result_matrix[, !("rId"), with = FALSE], rownames = rownames)
+  result_matrix <- as.matrix(result_matrix[, !("rId"), with = FALSE])
+  rownames(result_matrix) <- rownames
   
   # Replace 0 with NA to use na_col for red color
   result_matrix[result_matrix == 0] <- NA
@@ -131,9 +225,10 @@ heatmap_control_mapping_qc <- function(dt_treat,
                      color = grDevices::colorRampPalette(c("#CEEFC8", "#76d364"))(length(breaks) - 1),
                      breaks = breaks,
                      na_col = "red",
-                     main = "Mapping Counts Comparison between Treated and Control",
+                     main = "Counts of mapped controls",
                      cluster_cols = FALSE,
-                     angle_col = 45,
+                     cluster_rows = FALSE,
+                     angle_col = 90,
                      legend_breaks = unique_values,
                      legend_labels = unique_values)
 }